Extract formatMeetingWeekLine into util/timeFormat.ts
Moves the Week-range label logic out of CalendarNavbar.tsx into a
shared formatMeetingWeekLine (plus its monthNameForETDateString
helper), alongside formatMeetingDateLine which it mirrors: each
boundary's year is appended only when it differs from the current ET
year, rather than always showing the year like the inline version did.
